+++ qa/qa/checks/generic/all_caps_words.py
