--- qa/qa/checks/models/__init__.py
@@ -1,7 +1,0 @@
-"""Input and output models for QA checks."""
-
-from qa.checks.models.inputs import CheckData, SubmissionMetadata  # noqa
-from qa.checks.models.results import (
-    CheckResult,  # noqa
-    AlwaysPassContentCheckData,  # noqa
-)
